requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
h11==0.14.0
idna==3.4
msgpack==1.0.5
nats-py==2.2.0
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
